+++ src/main/java/main/arbitrage/infrastructure/exchange/upbit/dto/response/UpbitGetOrderResponse.java
@@ -16,7 +16,7 @@ public record UpbitGetOrderResponse(
     UpbitOrderEnums.State state,
     String createdAt,
     double volume,
-    double paidFee,
+    float paidFee,
     double executedVolume,
     List<Trade> trades) {
